refactor(components): migrate translations location to adjacent
