+++ src/lib/client/components/Footer/translation/en.ts
