+++ src/lib/client/components/Footer/translation/index.ts
