+++ src/lib/client/components/Header/translation/index.ts
