+++ src/lib/client/components/JobCard/translation/en.ts
