+++ src/lib/client/components/JobCard/translation/index.ts
