+++ src/lib/client/components/JobCard/translation/vi.ts
